Debug alerts not showing
