+++ src/App.jsx
@@ -1,5 +1,5 @@
 import "./App.css";
-import { useGlobalContext } from "./context";
+import { useAlertContext } from "./AlertContext";
 import Header from "./components/Header";
 import Footer from "./components/Footer";
 import Alerts from "./components/Alerts";
@@ -14,7 +14,7 @@ import MapContainer from "./components/MapContainer";
 // It will show up in the test text main, and also the developer console.
 
 function App() {
-  const { showModal, login } = useGlobalContext();
+  const { showModal, login } = useAlertContext();
 
   return (
     <main>
