+++ src/index.jsx
@@ -3,14 +3,17 @@ import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
-import { AppProvider } from "./context";
+import { PolygonProvider } from "./PolygonContext";
+import { AlertProvider } from "./AlertContext";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <React.StrictMode>
-    <AppProvider>
-      <App />
-    </AppProvider>
+    <AlertProvider>
+      <PolygonProvider>
+        <App />
+      </PolygonProvider>
+    </AlertProvider>
   </React.StrictMode>
 );
